Fix popup in  household manager
