+++ todoqueue_frontend/src/App.css
@@ -607,17 +607,10 @@ h1 {
   text-align: center;
   font-family: 'Arial', sans-serif;
   margin-top: 0px;
-  display: flex;
-}
-
-
-.household-container {
-  left: 50%;
-  transform: translateX(-50%);
-  position: absolute;
-  top: calc(50px + 2.5em);
+  display: block;
   width: 400px;
   font-family: 'Arial', sans-serif;
+  margin: 0 auto;
 }
 
 .household-wrapper {
@@ -628,6 +621,8 @@ h1 {
   border-radius: 4px;
   padding: 8px;
   margin-top: 10px;
+  margin-left: auto;
+  margin-right: auto;
 }
 
 .household-select {
