+++ todoqueue_frontend/src/components/spinner/Spinner.css
@@ -15,10 +15,26 @@
     border-style: solid;
     width: 10vh;
     height: 10vh;
-    animation: spin 2s linear infinite;
+    animation: scaleUp 0.75s forwards linear, spin 2s linear 0.75s infinite;
+    transform: scale(0);
+}
+
+@keyframes scaleUp {
+    0% {
+        transform: scale(0) rotate(0deg);
+    }
+
+    100% {
+        transform: scale(1) rotate(180deg);
+    }
 }
 
 @keyframes spin {
-    0% { transform: rotate(0deg); }
-    100% { transform: rotate(360deg); }
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(360deg);
+    }
 }
